requirements.txt
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
hyperliquid-python-sdk>=0.5.0
eth-account>=0.11.0
